# packages/backend/app/openapi.yaml
openapi: 3.0.3
info:
  title: FinMind API
  version: 0.1.0
  description: API for budgeting, bills, reminders and insights
servers:
  - url: http://localhost:8000
...
paths:
...
  /expenses:
    get:
      summary: List expenses
      tags: [Expenses]
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: List of expenses
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Expense'
              example:
                - { id: 10, amount: 12.5, currency: USD, category_id: 1, notes: Lunch, spent_at: 2025-08-10 }
                - { id: 11, amount: 49.0, currency: USD, category_id: 2, notes: "", spent_at: 2025-08-09 }
        '401':
          description: Unauthorized
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
    post:
      summary: Create expense
      tags: [Expenses]
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NewExpense'
            example:
              amount: 12.5
              currency: USD
              category_id: 1
              notes: Lunch
              spent_at: 2025-08-10
      responses:
        '201': { description: Created }
        '400':
          description: Validation error
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Error:
      type: object
      properties:
        error: { type: string }
      example: { error: "invalid credentials" }
...
    Expense:
      type: object
      properties:
        id: { type: integer }
        amount: { type: number, format: float }
        currency: { type: string }
        category_id: { type: integer }
        notes: { type: string, nullable: true }
        spent_at: { type: string, format: date }
    NewExpense:
      type: object
      required: [amount]
      properties:
        amount: { type: number, format: float }
        currency: { type: string, default: USD }
        category_id: { type: integer, nullable: true }
        notes: { type: string, nullable: true }
        spent_at: { type: string, format: date, nullable: true }
...